+++ st2actions/st2actions/query/mistral/v2.py
@@ -33,7 +33,7 @@ class MistralResultsQuerier(Querier):
         :type query_context: ``objext``
         :rtype: (``str``, ``object``)
         """
-        exec_id = query_context.get('mistral_execution_id', None)
+        exec_id = query_context.get('mistral', {}).get('execution_id', None)
         if not exec_id:
             raise Exception('Mistral execution id invalid in query_context %s.' %
                             str(query_context))