+++ st2actions/st2actions/runners/httprunner.py
@@ -80,7 +80,7 @@ class HttpRunner(ActionRunner):
         LOG.debug('action_parameters = %s', action_parameters)
         output = client.run()
         status = HttpRunner._get_result_status(output.get('status_code', None))
-        return (status, output)
+        return (status, output, None)
 
     def _get_http_client(self, action_parameters):
         body = action_parameters.get(ACTION_BODY, None)