+++ st2actions/tests/unit/test_mistral_v2.py
@@ -13,6 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import copy
 import json
 import uuid
 
@@ -71,6 +72,8 @@ PACK = 'generic'
 LOADER = FixturesLoader()
 FIXTURES = LOADER.load_fixtures(fixtures_pack=PACK, fixtures_dict=TEST_FIXTURES)
 
+MISTRAL_EXECUTION = {'id': str(uuid.uuid4()), 'state': 'RUNNING', 'workflow_name': None}
+
 # Workbook with a single workflow
 WB1_YAML_FILE_NAME = TEST_FIXTURES['workflows'][0]
 WB1_YAML_FILE_PATH = LOADER.get_fixture_file_path_abs(PACK, 'workflows', WB1_YAML_FILE_NAME)
@@ -79,6 +82,8 @@ WB1_YAML = yaml.safe_dump(WB1_SPEC, default_flow_style=False)
 WB1_NAME = '%s.%s' % (PACK, WB1_YAML_FILE_NAME.replace('.yaml', ''))
 WB1 = workbooks.Workbook(None, {'name': WB1_NAME, 'definition': WB1_YAML})
 WB1_OLD = workbooks.Workbook(None, {'name': WB1_NAME, 'definition': ''})
+WB1_EXEC = copy.deepcopy(MISTRAL_EXECUTION)
+WB1_EXEC['workflow_name'] = WB1_NAME
 
 # Workbook with many workflows
 WB2_YAML_FILE_NAME = TEST_FIXTURES['workflows'][1]
@@ -87,6 +92,8 @@ WB2_SPEC = FIXTURES['workflows'][WB2_YAML_FILE_NAME]
 WB2_YAML = yaml.safe_dump(WB2_SPEC, default_flow_style=False)
 WB2_NAME = '%s.%s' % (PACK, WB2_YAML_FILE_NAME.replace('.yaml', ''))
 WB2 = workbooks.Workbook(None, {'name': WB2_NAME, 'definition': WB2_YAML})
+WB2_EXEC = copy.deepcopy(MISTRAL_EXECUTION)
+WB2_EXEC['workflow_name'] = WB2_NAME
 
 # Workbook with many workflows but no default workflow is defined
 WB3_YAML_FILE_NAME = TEST_FIXTURES['workflows'][2]
@@ -95,6 +102,8 @@ WB3_SPEC = FIXTURES['workflows'][WB3_YAML_FILE_NAME]
 WB3_YAML = yaml.safe_dump(WB3_SPEC, default_flow_style=False)
 WB3_NAME = '%s.%s' % (PACK, WB3_YAML_FILE_NAME.replace('.yaml', ''))
 WB3 = workbooks.Workbook(None, {'name': WB3_NAME, 'definition': WB3_YAML})
+WB3_EXEC = copy.deepcopy(MISTRAL_EXECUTION)
+WB3_EXEC['workflow_name'] = WB3_NAME
 
 # Non-workbook with a single workflow
 WF1_YAML_FILE_NAME = TEST_FIXTURES['workflows'][3]
@@ -104,6 +113,8 @@ WF1_YAML = yaml.safe_dump(WF1_SPEC, default_flow_style=False)
 WF1_NAME = '%s.%s' % (PACK, WF1_YAML_FILE_NAME.replace('.yaml', ''))
 WF1 = workflows.Workflow(None, {'name': WF1_NAME, 'definition': WF1_YAML})
 WF1_OLD = workflows.Workflow(None, {'name': WF1_NAME, 'definition': ''})
+WF1_EXEC = copy.deepcopy(MISTRAL_EXECUTION)
+WF1_EXEC['workflow_name'] = WF1_NAME
 
 # Non-workbook with a many workflows
 WF2_YAML_FILE_NAME = TEST_FIXTURES['workflows'][4]
@@ -112,10 +123,11 @@ WF2_SPEC = FIXTURES['workflows'][WF2_YAML_FILE_NAME]
 WF2_YAML = yaml.safe_dump(WF2_SPEC, default_flow_style=False)
 WF2_NAME = '%s.%s' % (PACK, WF2_YAML_FILE_NAME.replace('.yaml', ''))
 WF2 = workflows.Workflow(None, {'name': WF2_NAME, 'definition': WF2_YAML})
+WF2_EXEC = copy.deepcopy(MISTRAL_EXECUTION)
+WF2_EXEC['workflow_name'] = WF2_NAME
 
 # Action executions' requirements
 ACTION_PARAMS = {'friend': 'Rocky'}
-EXECUTION = executions.Execution(None, {'id': str(uuid.uuid4()), 'state': 'RUNNING'})
 CHAMPION = worker.Worker(None)
 
 NON_EMPTY_RESULT = 'non-empty'
@@ -127,7 +139,7 @@ def process_create(payload):
 
 
 @mock.patch.object(LocalShellRunner, 'run', mock.
-                   MagicMock(return_value=(LIVEACTION_STATUS_SUCCEEDED, NON_EMPTY_RESULT)))
+                   MagicMock(return_value=(LIVEACTION_STATUS_SUCCEEDED, NON_EMPTY_RESULT, None)))
 @mock.patch.object(CUDPublisher, 'publish_create', mock.MagicMock(side_effect=process_create))
 @mock.patch.object(CUDPublisher, 'publish_update', mock.MagicMock(return_value=None))
 class TestMistralRunner(DbTestCase):
@@ -137,7 +149,7 @@ class TestMistralRunner(DbTestCase):
         super(TestMistralRunner, cls).setUpClass()
         runners_registrar.register_runner_types()
 
-        for name, fixture in six.iteritems(FIXTURES['actions']):
+        for _, fixture in six.iteritems(FIXTURES['actions']):
             instance = ActionAPI(**fixture)
             Action.add_or_update(ActionAPI.to_model(instance))
 
@@ -152,13 +164,18 @@ class TestMistralRunner(DbTestCase):
         mock.MagicMock(return_value=[WF1]))
     @mock.patch.object(
         executions.ExecutionManager, 'create',
-        mock.MagicMock(return_value=EXECUTION))
+        mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC)))
     def test_launch_workflow(self):
         MistralRunner.entry_point = mock.PropertyMock(return_value=WF1_YAML_FILE_PATH)
         execution = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS)
         execution = action_service.schedule(execution)
         execution = LiveAction.get_by_id(str(execution.id))
         self.assertEqual(execution.status, LIVEACTION_STATUS_RUNNING)
+
+        mistral_context = execution.context.get('mistral', None)
+        self.assertIsNotNone(mistral_context)
+        self.assertEqual(mistral_context['execution_id'], WF1_EXEC.get('id'))
+        self.assertEqual(mistral_context['workflow_name'], WF1_EXEC.get('workflow_name'))
 
     @mock.patch.object(
         workflows.WorkflowManager, 'list',
@@ -174,13 +191,18 @@ class TestMistralRunner(DbTestCase):
         mock.MagicMock(return_value=[WF1]))
     @mock.patch.object(
         executions.ExecutionManager, 'create',
-        mock.MagicMock(return_value=EXECUTION))
+        mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC)))
     def test_launch_when_workflow_definition_changed(self):
         MistralRunner.entry_point = mock.PropertyMock(return_value=WF1_YAML_FILE_PATH)
         execution = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS)
         execution = action_service.schedule(execution)
         execution = LiveAction.get_by_id(str(execution.id))
         self.assertEqual(execution.status, LIVEACTION_STATUS_RUNNING)
+
+        mistral_context = execution.context.get('mistral', None)
+        self.assertIsNotNone(mistral_context)
+        self.assertEqual(mistral_context['execution_id'], WF1_EXEC.get('id'))
+        self.assertEqual(mistral_context['workflow_name'], WF1_EXEC.get('workflow_name'))
 
     @mock.patch.object(
         workflows.WorkflowManager, 'list',
@@ -196,12 +218,17 @@ class TestMistralRunner(DbTestCase):
         mock.MagicMock(return_value=[WF1]))
     @mock.patch.object(
         executions.ExecutionManager, 'create',
-        mock.MagicMock(return_value=EXECUTION))
+        mock.MagicMock(return_value=executions.Execution(None, WF1_EXEC)))
     def test_launch_when_workflow_not_exists(self):
         execution = LiveActionDB(action=WF1_NAME, parameters=ACTION_PARAMS)
         execution = action_service.schedule(execution)
         execution = LiveAction.get_by_id(str(execution.id))
         self.assertEqual(execution.status, LIVEACTION_STATUS_RUNNING)
+
+        mistral_context = execution.context.get('mistral', None)
+        self.assertIsNotNone(mistral_context)
+        self.assertEqual(mistral_context['execution_id'], WF1_EXEC.get('id'))
+        self.assertEqual(mistral_context['workflow_name'], WF1_EXEC.get('workflow_name'))
 
     @mock.patch.object(
         workflows.WorkflowManager, 'list',
@@ -246,7 +273,7 @@ class TestMistralRunner(DbTestCase):
         mock.MagicMock(return_value=WB1))
     @mock.patch.object(
         executions.ExecutionManager, 'create',
-        mock.MagicMock(return_value=EXECUTION))
+        mock.MagicMock(return_value=executions.Execution(None, WB1_EXEC)))
     def test_launch_workbook(self):
         MistralRunner.entry_point = mock.PropertyMock(return_value=WB1_YAML_FILE_PATH)
         execution = LiveActionDB(action=WB1_NAME, parameters=ACTION_PARAMS)
@@ -254,6 +281,11 @@ class TestMistralRunner(DbTestCase):
         execution = LiveAction.get_by_id(str(execution.id))
         self.assertEqual(execution.status, LIVEACTION_STATUS_RUNNING)
 
+        mistral_context = execution.context.get('mistral', None)
+        self.assertIsNotNone(mistral_context)
+        self.assertEqual(mistral_context['execution_id'], WB1_EXEC.get('id'))
+        self.assertEqual(mistral_context['workflow_name'], WB1_EXEC.get('workflow_name'))
+
     @mock.patch.object(
         workflows.WorkflowManager, 'list',
         mock.MagicMock(return_value=[]))
@@ -268,13 +300,18 @@ class TestMistralRunner(DbTestCase):
         mock.MagicMock(return_value=WB2))
     @mock.patch.object(
         executions.ExecutionManager, 'create',
-        mock.MagicMock(return_value=EXECUTION))
+        mock.MagicMock(return_value=executions.Execution(None, WB2_EXEC)))
     def test_launch_workbook_with_many_workflows(self):
         MistralRunner.entry_point = mock.PropertyMock(return_value=WB2_YAML_FILE_PATH)
         execution = LiveActionDB(action=WB2_NAME, parameters=ACTION_PARAMS)
         execution = action_service.schedule(execution)
         execution = LiveAction.get_by_id(str(execution.id))
         self.assertEqual(execution.status, LIVEACTION_STATUS_RUNNING)
+
+        mistral_context = execution.context.get('mistral', None)
+        self.assertIsNotNone(mistral_context)
+        self.assertEqual(mistral_context['execution_id'], WB2_EXEC.get('id'))
+        self.assertEqual(mistral_context['workflow_name'], WB2_EXEC.get('workflow_name'))
 
     @mock.patch.object(
         workflows.WorkflowManager, 'list',
@@ -290,7 +327,7 @@ class TestMistralRunner(DbTestCase):
         mock.MagicMock(return_value=WB3))
     @mock.patch.object(
         executions.ExecutionManager, 'create',
-        mock.MagicMock(return_value=EXECUTION))
+        mock.MagicMock(return_value=executions.Execution(None, WB3_EXEC)))
     def test_launch_workbook_with_many_workflows_no_default(self):
         MistralRunner.entry_point = mock.PropertyMock(return_value=WB3_YAML_FILE_PATH)
         execution = LiveActionDB(action=WB3_NAME, parameters=ACTION_PARAMS)
@@ -313,13 +350,18 @@ class TestMistralRunner(DbTestCase):
         mock.MagicMock(return_value=WB1))
     @mock.patch.object(
         executions.ExecutionManager, 'create',
-        mock.MagicMock(return_value=EXECUTION))
+        mock.MagicMock(return_value=executions.Execution(None, WB1_EXEC)))
     def test_launch_when_workbook_definition_changed(self):
         MistralRunner.entry_point = mock.PropertyMock(return_value=WB1_YAML_FILE_PATH)
         execution = LiveActionDB(action=WB1_NAME, parameters=ACTION_PARAMS)
         execution = action_service.schedule(execution)
         execution = LiveAction.get_by_id(str(execution.id))
         self.assertEqual(execution.status, LIVEACTION_STATUS_RUNNING)
+
+        mistral_context = execution.context.get('mistral', None)
+        self.assertIsNotNone(mistral_context)
+        self.assertEqual(mistral_context['execution_id'], WB1_EXEC.get('id'))
+        self.assertEqual(mistral_context['workflow_name'], WB1_EXEC.get('workflow_name'))
 
     @mock.patch.object(
         workflows.WorkflowManager, 'list',
@@ -335,12 +377,17 @@ class TestMistralRunner(DbTestCase):
         mock.MagicMock(return_value=WB1))
     @mock.patch.object(
         executions.ExecutionManager, 'create',
-        mock.MagicMock(return_value=EXECUTION))
+        mock.MagicMock(return_value=executions.Execution(None, WB1_EXEC)))
     def test_launch_when_workbook_not_exists(self):
         execution = LiveActionDB(action=WB1_NAME, parameters=ACTION_PARAMS)
         execution = action_service.schedule(execution)
         execution = LiveAction.get_by_id(str(execution.id))
         self.assertEqual(execution.status, LIVEACTION_STATUS_RUNNING)
+
+        mistral_context = execution.context.get('mistral', None)
+        self.assertIsNotNone(mistral_context)
+        self.assertEqual(mistral_context['execution_id'], WB1_EXEC.get('id'))
+        self.assertEqual(mistral_context['workflow_name'], WB1_EXEC.get('workflow_name'))
 
     @mock.patch.object(
         workflows.WorkflowManager, 'list',